Fix up basic auth
Previously, for selfish reasons, we weren't using base64 in our auth setup
which made the gravel gateway incompatible with standard http libs.

This commit updates the basic auth handler such that it does base64 decode
things when checking passwords.

For backwards compatibility reasons, if we fail to base64 decode the header,
or if its invalid base64, or if it doesn't contain a : (i.e. its an invalid
Auth header), then we fall back to treating it as plain text to try
and make this as less of a breaking change as possible